Issue ?? - Cameron James Updated unit tests to get more coverage, also added logging so the soap ws messages are put into a separate log file.

+++ trunk/0.3/Apromore-Core/apromore-core/src/main/java/org/apromore/dao/CanonicalDao.java
@@ -18,7 +18,7 @@ public interface CanonicalDao {
      * Find the canoncial record by it's process id.
      * @return the canonical process, a list of them for all the different versions.
      */
-    List<Canonical> findByProcessId(final int processId);
+    List<Canonical> findByProcessId(final long processId);
